Update AutoMod features and Add new status

+++ Cogs/AutoFeatures.py
@@ -16,16 +16,27 @@ class AutoFeatures(commands.Cog):
     async def on_member_join(self, member):
         role = discord.utils.get(member.guild.roles, name="Member")
         await member.add_roles(role)
-        channel = discord.utils.get(member.guild.text_channels, name='║・𝐖𝐞𝐥𝐜𝐨𝐦𝐞')
-        await channel.send(f"Welcome to the server, {member.mention}! 🎉 You've been assigned the **Member** role.")
+        channel = discord.utils.get(member.guild.text_channels, name='welcome')
+        await channel.send(f"Welcome to the server, {member.mention}! 🎉 You've been assigned the **Member** role. Enjoy your stay! <3")
         
     # Bad Words Filter (AutoMod)
     @commands.Cog.listener()
     async def on_message(self, message):
-        bad_words = ["nigger", "nigga", "fuck", "bitch", "hoe",]
+        if message.author == self.bot.user:
+            return
+
+        bad_words = ["nigga", "nga", "nigger", "fuck", "slut"] #will test when im home. at the time of typing this 9:47am. And Will add more later.
+        
         if any(word in message.content.lower() for word in bad_words):
-            await message.delete()
-            await message.channel.send(f"{message.author.mention}, watch your language! No bad words!")
+            try:
+                await message.delete()
+                await message.channel.send(
+                    f"{message.author.mention}, watch your language!",
+                    delete_after=5
+                )
+            except discord.Forbidden:
+                print("Bot lacks permission to delete messages. Please update the bot perms via role or Contact the maker/owner of the bot to fix via Discord Dev Portal.")
+
         await self.bot.process_commands(message)
 
 def setup(bot):
